% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaner.R
\name{data_cleaner}
\alias{data_cleaner}
\title{Clean web-scraped data.}
\usage{
data_cleaner(scraped_df)
}
\arguments{
\item{scraped_df}{A tibble containing web-scraped data like listing url, price and house type.}
}
\value{
tibble A cleaned tibble containing information like listing url, price, number of bedrooms, area in sqft, and city.
}
\description{
A function to clean web-scraped data given the tibble containing raw data .
}
