% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filter.R
\name{data_filter}
\alias{data_filter}
\title{Filter a cleaned tibble based on user input}
\usage{
data_filter(df, min_price, max_price, sqrt_ft, num_bedroom, city)
}
\arguments{
\item{df}{tibble}

\item{min_price}{number}

\item{max_price}{number}

\item{sqrt_ft}{number}

\item{num_bedroom}{int}

\item{city}{char}
}
\value{
tibble
}
\description{
Filter a cleaned tibble based on user input
}
