% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polite-scrape.R
\name{polite_fetch_rtxt}
\alias{polite_fetch_rtxt}
\title{function to get robots.txt is structured form. Memoised}
\usage{
polite_fetch_rtxt(..., user_agent, delay, verbose)
}
\description{
function to get robots.txt is structured form. Memoised
}
