% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polite-scrape.R
\name{polite_read_html}
\alias{polite_read_html}
\title{function that actually fetches response from the web}
\usage{
polite_read_html(
  url,
  ...,
  delay = 5,
  user_agent = paste0("polite ", getOption("HTTPUserAgent"), "bot"),
  force = FALSE,
  verbose = FALSE
)
}
\description{
function that actually fetches response from the web
}
