% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{scraper}
\alias{scraper}
\title{Function to scrape housing data from a given craiglist url}
\usage{
scraper(url, online = False)
}
\arguments{
\item{url}{The given craiglist url to scrape the data from}

\item{online}{Whether the data is scraped directly online from the url (default = False). False means the data is scraped from local HTML file}
}
\value{
tibble A dataframe containing listing information such as price, house type and neighborhood.
}
\description{
Function to scrape housing data from a given craiglist url
}
\examples{
scraper(url = 'https://vancouver.craigslist.org/search/van/apa')
}
