% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaner.R
\name{data_cleaner}
\alias{data_cleaner}
\title{Clean web-scraped data.}
\usage{
data_cleaner(scraped_df)
}
\arguments{
\item{scraped_df}{A tibble containing web-scraped data like listing url, price and house type.}
}
\value{
tibble A cleaned tibble containing information like listing url, price, number of bedrooms, area in sqft, and city.
}
\description{
A function to clean web-scraped data given the tibble containing raw data .
}
\examples{
data_cleaner(tibble::tibble(“price”= c($1,300", $1,200"), “house_type”= c(“1br-600ft2-“, “2br-“), “listing_url” = c(“https://vancouver.craigslist.org/bnc/apa/d/burnaby-must-see-1br-suite/7282955370.html”, “https://vancouver.craigslist.org/rds/apa/d/surrey-bedroom-basement-for-rent/7273877138.html”)))
}
