% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{scraper}
\alias{scraper}
\title{Function to scrape housing data from a given craiglist url}
\usage{
scraper(url, online = FALSE)
}
\arguments{
\item{url}{The given Craiglist url to scrape the data from.}

\item{online}{Whether the data is scraped directly from the url online (default = False).
False means the data is scraped from a local HTML file.}
}
\value{
tibble A tibble containing listing information such as listing url, price, and house type
}
\description{
Function to scrape housing data from a given craiglist url
}
\examples{
scraper(url = "https://vancouver.craigslist.org/d/apartments-housing-for-rent/search/apa", online = FALSE)
}
