% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_email.R
\name{send_email}
\alias{send_email}
\title{A function to email search results.}
\usage{
send_email(
  email_recipient,
  filtered_data,
  email_subject = "Results from RHouseHunter"
)
}
\arguments{
\item{email_recipient}{A string containing the email address of the recipient.}

\item{filtered_data}{A tibble containing the filtered search results.}

\item{email_subject}{A string containing the desired subject of the email.
Optional argument with default value of  'Results from RHouseHunter'.}
}
\value{
None.
}
\description{
A function to email search results.
}
\examples{
send_email("helloworld@gmail.com", df)
send_email("helloworld@gmail.com", df, "Scraping Results 2")
}
